requirement.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tokenizers>=0.13.0
transformers>=4.20.0
tqdm>=4.60.0
requests>=2.26.0
Pillow>=8.3.0
packaging>=21.0
